requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
pandas>=2.0.0
plotly>=5.17.0
numpy>=1.25.2,<2.6.0
pydantic>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
xgboost>=2.0.0
numba>=0.58.0
shap>=0.48.0
geopy>=2.3.0
scipy>=1.10.0
statsmodels>=0.14.0
lime>=0.2.0
lifelines>=0.27.0
networkx>=3.0
prophet>=1.1.0
tensorflow>=2.13.0
torch>=2.0.0
umap-learn>=0.5.3
imbalanced-learn>=0.11.0
